requirements.txt
django
djangorestframework
requests
urllib3
django-nose
coverage
behave
discord